Title: 사각형 채우기 3, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250924/사각형 채우기 3/rectangle-fill-3.cpp
@@ -7,7 +7,7 @@ int n;
 int main() {
     cin >> n;
 
-    int rect[1001]{};
+    long long int rect[1001]{};
 
     rect[0] = 1;
     rect[1] = 2;
